accessible/tests/mochitest/text/test_whitespaces.html

<!DOCTYPE html>
<html>
<head>
  <title>getText... methods tests on string with whitespaces for plain text containers</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="../common.js"></script>

  <script type="application/javascript"
          src="../text.js"></script>
  <script type="application/javascript">
    function doTest()
    {
      // __B__r__a__v__e__ __S__i__r__ __ __R__o__b__i__n__ __ __ __r__a__n
      //  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21

      ////////////////////////////////////////////////////////////////////////
      // getTextAfterOffset

      var IDs = [ "input", "div", "editable", "textarea" ];

      // BOUNDARY_WORD_START
      testTextAfterOffset(0, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextAfterOffset(5, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextAfterOffset(6, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);
      testTextAfterOffset(9, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);
      testTextAfterOffset(10, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);
      testTextAfterOffset(11, BOUNDARY_WORD_START, "ran", 19, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(16, BOUNDARY_WORD_START, "ran", 19, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(18, BOUNDARY_WORD_START, "ran", 19, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(19, BOUNDARY_WORD_START, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kOk, kTodo, kTodo);

      // BOUNDARY_WORD_END
      testTextAfterOffset(0, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAfterOffset(4, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAfterOffset(5, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAfterOffset(6, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAfterOffset(8, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAfterOffset(9, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAfterOffset(10, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAfterOffset(11, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAfterOffset(15, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAfterOffset(16, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAfterOffset(17, BOUNDARY_WORD_END, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(18, BOUNDARY_WORD_END, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(19, BOUNDARY_WORD_END, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(21, BOUNDARY_WORD_END, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);
      testTextAfterOffset(22, BOUNDARY_WORD_END, "", 22, 22,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kOk, kTodo, kTodo);

      ////////////////////////////////////////////////////////////////////////
      // getTextBeforeOffset

      var IDs = [ "input", "div", "editable", "textarea" ];

      // BOUNDARY_WORD_START
      testTextBeforeOffset(0, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(5, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(6, BOUNDARY_WORD_START, "Brave ", 0, 6, IDs);
      testTextBeforeOffset(9, BOUNDARY_WORD_START, "Brave ", 0, 6, IDs);
      testTextBeforeOffset(10, BOUNDARY_WORD_START, "Brave ", 0, 6, IDs);
      testTextBeforeOffset(11, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextBeforeOffset(15, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextBeforeOffset(16, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextBeforeOffset(17, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextBeforeOffset(18, BOUNDARY_WORD_START, "Sir  ", 6, 11, IDs);
      testTextBeforeOffset(19, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);
      testTextBeforeOffset(20, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);
      testTextBeforeOffset(21, BOUNDARY_WORD_START, "Robin   ", 11, 19, IDs);

      // BOUNDARY_WORD_END
      testTextBeforeOffset(0, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(4, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(5, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(6, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextBeforeOffset(7, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextBeforeOffset(8, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextBeforeOffset(9, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextBeforeOffset(10, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextBeforeOffset(11, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextBeforeOffset(15, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextBeforeOffset(16, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextBeforeOffset(17, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextBeforeOffset(18, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextBeforeOffset(19, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextBeforeOffset(21, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextBeforeOffset(22, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);

      ////////////////////////////////////////////////////////////////////////
      // getTextAtOffset

      // BOUNDARY_WORD_START
      testTextAtOffset(0, BOUNDARY_WORD_START, "Brave ", 0, 6,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(5, BOUNDARY_WORD_START, "Brave ", 0, 6,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(6, BOUNDARY_WORD_START, "Sir  ", 6, 11,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(8, BOUNDARY_WORD_START, "Sir  ", 6, 11,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(9, BOUNDARY_WORD_START, "Sir  ", 6, 11,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(10, BOUNDARY_WORD_START, "Sir  ", 6, 11,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(11, BOUNDARY_WORD_START, "Robin   ", 11, 19,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(15, BOUNDARY_WORD_START, "Robin   ", 11, 19,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(16, BOUNDARY_WORD_START, "Robin   ", 11, 19,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(17, BOUNDARY_WORD_START, "Robin   ", 11, 19,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(18, BOUNDARY_WORD_START, "Robin   ", 11, 19,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kOk, kOk, kOk);
      testTextAtOffset(19, BOUNDARY_WORD_START, "ran", 19, 22,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);
      testTextAtOffset(21, BOUNDARY_WORD_START, "ran", 19, 22,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);
      testTextAtOffset(22, BOUNDARY_WORD_START, "ran", 19, 22,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);

      // BOUNDARY_WORD_END
      testTextAtOffset(0, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextAtOffset(4, BOUNDARY_WORD_END, "Brave", 0, 5, IDs);
      testTextAtOffset(5, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAtOffset(6, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAtOffset(8, BOUNDARY_WORD_END, " Sir", 5, 9, IDs);
      testTextAtOffset(9, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAtOffset(10, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAtOffset(11, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAtOffset(15, BOUNDARY_WORD_END, "  Robin", 9, 16, IDs);
      testTextAtOffset(16, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(17, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(18, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(19, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(20, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(21, BOUNDARY_WORD_END, "   ran", 16, 22, IDs);
      testTextAtOffset(22, BOUNDARY_WORD_END, "   ran", 16, 22, ["input", "div", "editable"]);
      testTextAtOffset(22, BOUNDARY_WORD_END, "   ran\n", 16, 23, [ "textarea" ]);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="getText... methods tests on string with whitespaces for plain text containers"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=610568">
    Bug 610568
  </a>
  <a target="_blank"
     title="getTextAtOffset for word boundaries: beginning of a new life"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=853340">
    Bug 853340
  </a>
  <a target="_blank"
     title="getTextBeforeOffset for word boundaries: evolving"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=855732">
    Bug 855732
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  <input id="input" value="Brave Sir  Robin   ran"/>
  <div id="div">Brave Sir  Robin   ran</div>
  <div id="editable" contenteditable="true">Brave Sir  Robin   ran</div>
  <textarea id="textarea" cols="300">Brave Sir  Robin   ran</textarea>
  </pre>

</body>
</html>
